backend/app/models/department.py
from typing import List

from sqlalchemy import String
from sqlalchemy.orm import Mapped, mapped_column, relationship
from typing_extensions import TYPE_CHECKING

from ..database import Base

if TYPE_CHECKING:
    from .employee import Employee


class Department(Base):
    __tablename__ = "departments"
    id: Mapped[int] = mapped_column(primary_key=True)
    name: Mapped[str] = mapped_column(String(50), nullable=False, unique=True)
    description: Mapped[str] = mapped_column(String(200))

    employees: Mapped[List[Employee]] = relationship(back_populates="department")
