backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi>=0.135.2",
    "psycopg2>=2.9.11",
    "pwdlib[argon2]>=0.3.0",
    "pydantic[email]>=2.12.5",
    "pydantic-settings>=2.13.1",
    "pyjwt>=2.12.1",
    "sqlalchemy>=2.0.48",
    "uvicorn>=0.42.0",
    "python-multipart>=0.0.22",
]

[dependency-groups]
test = [
    "pytest>=8.0",
    "httpx>=0.27",
]
